+++ R/data_dacs.R
@@ -1,0 +1,166 @@
+#' Software failure data
+#'
+#' Software failure data (time data, grouped data) collected by John Musa and Yoshihiro Tohma
+#'
+#' The follwoing data are software failure data in real application, which was
+#' collected by John Musa. The name without \code{g} (e.g. \code{sys1}, \code{sys2}) indicates
+#' the interval time of failures in CPU time, and the last negative value
+#' means the time in which there is no failure from the last failure. The name with
+#' the suffix \code{g} (e.g. \code{sys1g}, \code{sys2g}) means the grouped fauilure data, i.e., the sequence
+#' of the number of failures per woring day.
+#' \describe{
+#'  \item{sys1, sys1g}{136 failures have occured during system test and operations in the program with 21,700 instructions.}
+#'  \item{sys2, sys2g}{54 failures have occured during system test and operations in the program with 27,700 instructions.}
+#'  \item{sys3, sys3g}{38 failures have occured during system test and operations in the program with 23,400 instructions.}
+#'  \item{sys4, sys4g}{53 failures have occured during system test and operations in the program with 33,500 instructions.}
+#'  \item{sys5, sys5g}{831 failures have occured during system test in the program with 2,445,000 instructions.}
+#'  \item{sys6, sys6g}{73 failures have occured during subsystem test in the program with 5,700 instructions.}
+#'  \item{sys14c, sys14cg}{36 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{sys17, sys17g}{38 failures have occured during system test in the program with 61,900 instructions.}
+#'  \item{sys27, sys27g}{41 failures have occured during system test in the program with 126,100 instructions.}
+#'  \item{sys40, sys40g}{101 failures have occured during system test in the program with 180,000 instructions.}
+#'  \item{ss1a, ss1ag}{112 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{ss1b, ss1bg}{375 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{ss1c, ss1cg}{277 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{ss2, ss2g}{192 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{ss3, ss3g}{278 failures have occured during operations in the program with hundreds of thousands instructions.}
+#'  \item{ss4, ss4g}{196 failures have occured during operations in the program with hundreds of thousands instructions.}
+#' }
+#' The data \code{tohma} is the faliure data collected by Yoshihiro Tohma. This is the grouped failure data consisting of
+#' the number of failures per a test.
+#'
+#' @name dacs
+#' @docType data
+NULL
+
+#' @name sys1
+#' @rdname dacs
+NULL
+
+#' @name sys2
+#' @rdname dacs
+NULL
+
+#' @name sys3
+#' @rdname dacs
+NULL
+
+#' @name sys4
+#' @rdname dacs
+NULL
+
+#' @name sys5
+#' @rdname dacs
+NULL
+
+#' @name sys6
+#' @rdname dacs
+NULL
+
+#' @name sys14c
+#' @rdname dacs
+NULL
+
+#' @name sys17
+#' @rdname dacs
+NULL
+
+#' @name sys27
+#' @rdname dacs
+NULL
+
+#' @name sys40
+#' @rdname dacs
+NULL
+
+#' @name ss1a
+#' @rdname dacs
+NULL
+
+#' @name ss1b
+#' @rdname dacs
+NULL
+
+#' @name ss1c
+#' @rdname dacs
+NULL
+
+#' @name ss2
+#' @rdname dacs
+NULL
+
+#' @name ss3
+#' @rdname dacs
+NULL
+
+#' @name ss4
+#' @rdname dacs
+NULL
+
+#' @name sys1g
+#' @rdname dacs
+NULL
+
+#' @name sys2g
+#' @rdname dacs
+NULL
+
+#' @name sys3g
+#' @rdname dacs
+NULL
+
+#' @name sys4g
+#' @rdname dacs
+NULL
+
+#' @name sys5g
+#' @rdname dacs
+NULL
+
+#' @name sys6g
+#' @rdname dacs
+NULL
+
+#' @name sys14cg
+#' @rdname dacs
+NULL
+
+#' @name sys17g
+#' @rdname dacs
+NULL
+
+#' @name sys27g
+#' @rdname dacs
+NULL
+
+#' @name sys40g
+#' @rdname dacs
+NULL
+
+#' @name ss1ag
+#' @rdname dacs
+NULL
+
+#' @name ss1bg
+#' @rdname dacs
+NULL
+
+#' @name ss1cg
+#' @rdname dacs
+NULL
+
+#' @name ss2g
+#' @rdname dacs
+NULL
+
+#' @name ss3g
+#' @rdname dacs
+NULL
+
+#' @name ss4g
+#' @rdname dacs
+NULL
+
+#' @name tohma
+#' @rdname dacs
+NULL