+++ R/dataform_nhpp.R
@@ -10,6 +10,7 @@
 #' @param type Either 0 or 1. If 1, a fault is detected just at the end of corresponding time interval.
 #' This is used to represent the fault time data. If 0, no fault is detected at the end of interval.
 #' @param te A numeric value for the time interval from the last fault to the observation time.
+#' @param data A dataframe. The arguments; time, fault, type, te can also be selected as the columns of dataframe.
 #' @return A list with the attribute class='Rsrat.faultdata';
 #' \item{time}{A vector for time interval.}
 #' \item{fault}{A vector for the number of detected faults in intervals.}
@@ -19,28 +20,62 @@
 #' \item{mean}{A numeric value for mean fault detection time from data.}
 #' \item{max}{A numeric value for maximum of fault detection time.}
 #' @examples
+#' data(tomcat5)
 #' faultdata(time=c(1,1,1,1), fault=c(0,1,0,5))
 #' faultdata(time=c(3,1,7,15,12), te=3)
+#' faultdata(time=time, fault=fault, data=tomcat5.catalina)
 #' @export
 
-faultdata <- function(time, fault, type, te) {
+faultdata <- function(time, fault, type, te, data = data.frame()) {
+  if (!missing(time)) {
+    e <- substitute(time)
+    time <- eval(e, data, parent.frame())
+  }
+  else {
+    time <- NULL
+  }
+  if (!missing(fault)) {
+    e <- substitute(fault)
+    fault <- eval(e, data, parent.frame())
+  }
+  else {
+    fault <- NULL
+  }
+  if (!missing(type)) {
+    e <- substitute(type)
+    type <- eval(e, data, parent.frame())
+  }
+  else {
+    type <- NULL
+  }
+  if (!missing(te)) {
+    e <- substitute(te)
+    te <- eval(e, data, parent.frame())
+  }
+  else {
+    te <- NULL
+  }
+  list(time, fault, type, te)
+  .faultdata.nhpp(time, fault, type, te)
+}
 
-  if (missing(time)) {
-    if (missing(fault)) {
+.faultdata.nhpp <- function(time, fault, type, te) {
+  if (is.null(time)) {
+    if (is.null(fault)) {
       stop("Invalid data: Either time or fault is required.")
     } else {
-      if (!missing(te)) {
+      if (!is.null(te)) {
         warning("te is ignored.")
       }
-      if (missing(type)) {
+      if (is.null(type)) {
         type <- rep.int(0L, length(fault))
       }
       time <- rep.int(1L, length(fault))
     }
   } else {
-    if (missing(fault)) {
-      if (missing(type)) {
-        if (missing(te)) {
+    if (is.null(fault)) {
+      if (is.null(type)) {
+        if (is.null(te)) {
           stop("Invalid data: Either type or te is required when fault is missing.")
         } else {
           type <- rep.int(1L, length(time))
@@ -49,16 +84,16 @@ faultdata <- function(time, fault, type, te) {
           fault <- rep.int(0L, length(time))
         }
       } else {
-        if (!missing(te)) {
+        if (!is.null(te)) {
           warning("te is ignored.")
         }
         fault <- rep.int(0L, length(time))
       }
     } else {
-      if (!missing(te)) {
+      if (!is.null(te)) {
         warning("te is ignored.")
       }
-      if (missing(type)) {
+      if (is.null(type)) {
         type <- rep.int(0L, length(time))
       }
     }